pyproject.toml
[project]
name = "hoopla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "nltk==3.9.1",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "sentence-transformers>=5.1.1",
]
